Add fixtures for Pyomop and metadata in tests
