+++ tests/conftest.py
@@ -7,4 +7,16 @@
     https://pytest.org/latest/plugins.html
 """
 
-# import pytest
+import asyncio
+import pytest
+
+@pytest.fixture
+def pyomop_fixture():
+    from src.pyomop import CdmEngineFactory
+    cdm = CdmEngineFactory()
+    return cdm
+
+@pytest.fixture
+def metadata_fixture():
+    from src.pyomop import metadata
+    return metadata
