+++ tests/test_a_pyomop.py
@@ -1,17 +1,6 @@
 import asyncio
 import pytest
 
-@pytest.fixture
-def pyomop_fixture():
-    from src.pyomop import CdmEngineFactory
-    cdm = CdmEngineFactory()
-    return cdm
-
-@pytest.fixture
-def metadata_fixture():
-    from src.pyomop import metadata
-    return metadata
-
 @staticmethod
 @pytest.mark.asyncio
 def test_create_db(pyomop_fixture, metadata_fixture, capsys):
